Release 2 0 1

Approved-by: Arnaud Hours
Approved-by: Alexis Hermanns

+++ src/Storefront/Controller/LengowCronController.php
@@ -182,7 +182,7 @@ class LengowCronController extends LengowAbstractFrontController
                 ? $request->query->get(LengowImport::PARAM_DEBUG_MODE) === '1'
                 : null,
             LengowImport::PARAM_LOG_OUTPUT => $request->query->get(LengowImport::PARAM_LOG_OUTPUT) === '1',
-            LengowImport::PARAM_DAYS => (int) $request->query->get(LengowImport::PARAM_DAYS),
+            LengowImport::PARAM_DAYS => (float) $request->query->get(LengowImport::PARAM_DAYS),
             LengowImport::PARAM_CREATED_FROM => $request->query->get(LengowImport::PARAM_CREATED_FROM),
             LengowImport::PARAM_CREATED_TO => $request->query->get(LengowImport::PARAM_CREATED_TO),
             LengowImport::PARAM_LIMIT => (int) $request->query->get(LengowImport::PARAM_LIMIT),
